server/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose[cryptography]
bcrypt
python-dotenv
pydantic[email]
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
jinja2